put all the parser.cpp classes in parser.h, so i can use it in the interpreter file, created interpreter file.
