+++ interpreter/bin/compile.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-g++ -std=c++17 -Wall -Wextra ../src/lexer.cpp ../src/parser.cpp ../src/commons.cpp ../src/roscript.cpp -o ros
+g++ -std=c++17 -Wall -Wextra ../src/lexer.cpp ../src/parser.cpp ../src/commons.cpp ../src/interpreter.cpp ../src/roscript.cpp -o ros
