+++ interpreter/src/interpreter.h
@@ -1,0 +1,2 @@
+#include "parser.h"
+void interpret(vector<ASTNode*> AST);
